requirements.txt
# ===============================
#  應用程式依賴 (Application Dependencies)
# ===============================

# Web 框架與伺服器
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# WebRTC 與 WebSocket 支援 
websockets>=12.0  # FastAPI 處理 WebSocket 連線的底層依賴
aiortc>=1.5.0     # 實現 WebRTC 信令伺服器及後端處理的關鍵函式庫

# API 客戶端
openai==1.3.7
httpx>=0.25.2

# 音訊處理
pydub==0.25.1

# 工具
python-dotenv==1.0.0
pydantic==2.5.0

# ===============================
#  開發與測試依賴 (Development & Testing Dependencies)
# ===============================

# 測試框架
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# 程式碼格式化工具
black==23.11.0